third_party/blink/web_tests/virtual/threaded/compositing/img-element-transform.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<body onload="runTest()">
  <img id="target" src="../../../animations/resources/green-100.png">
</body>

<script>
  function runTest() {
    promise_test(async(t) => {
      let animation = target.animate({transform: ['rotate(0deg)', 'rotate(180deg)']}, 100000);
        return animation.ready.then(() => {
          assert_true(internals.isCompositedAnimation(animation));
        });
    }, '<img> elements should run compositor animations');
  }
</script>
